+++ hyperspyui/tests/test_load.py
@@ -13,3 +13,8 @@ def test_load(mainwindow, tmp_path):
     assert mainwindow.get_selected_signal() is None
     mainwindow.load([str(filename)])
     assert mainwindow.get_selected_signal() is not None
+
+
+def test_get_accepted_extensions(mainwindow):
+    extensions = mainwindow.get_accepted_extensions()
+    assert 'dm3' in extensions